+++ middleware/linkkit/sdk-c/src/services/ota/iotx_ota.c
@@ -743,7 +743,7 @@ int IOT_OTA_FetchYield(void *handle, char *buf, uint32_t buf_len, uint32_t timeo
 }
 
 
-int IOT_OTA_Ioctl(void *handle, IOT_OTA_CmdType_t type, void *buf, size_t buf_len)
+int IOT_OTA_Ioctl(void *handle, IOT_OTA_CmdType_t type, void *buf, int buf_len)
 {
     OTA_Struct_pt h_ota = (OTA_Struct_pt) handle;
